LATIR-6: Adding reusable Link component
