+++ package.json
@@ -5,6 +5,7 @@
   "homepage": "fundacionlatir.org",
   "dependencies": {
     "@juan-utils/functions": "^2.3.0",
+    "prop-types": "^15.7.2",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-router": "^5.1.2",
